Moved tokens to separate file.
